fix overflow

+++ conch/render/RenderDriver/OpenGLESDriver/RenderDevice/GLESVertexBuffer.h
@@ -26,9 +26,9 @@ class GLESVertexBuffer // : public VertexBuffer
     void setDataLength(int byteLength);
 
     void setData(const char *buffer, int bufferOffset /* = 0*/, int dataStartIndex /* = 0*/,
-                 int dataCount /*= Number.MAX_SAFE_INTEGER*/);
+        double dataCount /*= Number.MAX_SAFE_INTEGER*/);
     void setDataJS(JSValueAsParam data, int bufferOffset /* = 0*/, int dataStartIndex /* = 0*/,
-                   int dataCount /*= Number.MAX_SAFE_INTEGER*/);
+        double dataCount /*= Number.MAX_SAFE_INTEGER*/);
     void bind();
     void unbind();
     void orphanStorage();
